Code splitted in separate modules. Completed filter reset.
