+++ source/js/index.js
@@ -1,6 +1,8 @@
 import '../styles/style.scss';
-import './js_module/indexModule.js';
 import {hamburgerFunctionality, siteSearchFunctionality} from './js_module/headerSiteFunctionality.js';
+import {arrowSwitchSlider} from './js_module/arrowSwitchSlider.js';
 
-hamburgerFunctionality.hamburgerMenu.addEventListener('click', hamburgerFunctionality.toggleHamburgerMenu);
-siteSearchFunctionality.searchButton.addEventListener('click', siteSearchFunctionality.toggleSearchInput);
+hamburgerFunctionality.hamburgerMenuSelector.addEventListener('click', hamburgerFunctionality.toggleHamburgerMenu);
+siteSearchFunctionality.searchButtonSelector.addEventListener('click', siteSearchFunctionality.toggleSearchInput);
+
+arrowSwitchSlider.mainSectionSelector.addEventListener('click', arrowSwitchSlider.switchArrowHandler);
